services/wristband-simulator/requirements.txt
requests   
paho-mqtt